Clean tests up a little bit
