+++ ee/clickhouse/queries/funnels/test/test_funnel_trends_new.py
@@ -1,14 +1,12 @@
-from datetime import date, datetime, timedelta
-from typing import Union, cast
+from datetime import datetime
+from typing import cast
 from uuid import uuid4
 
-from ee.clickhouse.client import format_sql, sync_execute
 from ee.clickhouse.models.event import create_event
 from ee.clickhouse.queries.funnels.funnel_trends import ClickhouseFunnelTrendsNew
 from ee.clickhouse.util import ClickhouseTestMixin
 from posthog.constants import INSIGHT_FUNNELS, TRENDS_LINEAR
 from posthog.models.filters import Filter
-from posthog.models.filters.mixins.funnel_window_days import FunnelWindowDaysMixin
 from posthog.models.person import Person
 from posthog.test.base import APIBaseTest
 
@@ -36,25 +34,81 @@ def _create_event(**kwargs):
     create_event(**kwargs)
 
 
-class TestFunnelTrends(ClickhouseTestMixin, APIBaseTest):
+class TestFunnelTrendsNew(ClickhouseTestMixin, APIBaseTest):
     maxDiff = None
 
-    def test_returns(self):
+    def test_no_event_in_period(self):
+        _create_person(distinct_ids=[USER_A_DISTINCT_ID], team=self.team)
+
+        _create_event(event=STEP_1_EVENT, distinct_id=USER_A_DISTINCT_ID, team=self.team, timestamp=TIME_0)
+
         filter = Filter(
             data={
                 "insight": INSIGHT_FUNNELS,
                 "display": TRENDS_LINEAR,
-                "interval": "hour",
-                "date_from": "2021-05-01 00:00:00",
-                "date_to": "2021-05-07 00:00:00",
+                "interval": "day",
+                "date_from": TIME_1,
+                "date_to": TIME_99,
                 "funnel_window_days": 7,
                 "events": [
-                    {"id": "step one", "order": 0},
-                    {"id": "step two", "order": 1},
-                    {"id": "step three", "order": 2},
+                    {"id": STEP_1_EVENT, "order": 0},
+                    {"id": STEP_2_EVENT, "order": 1},
+                    {"id": STEP_3_EVENT, "order": 2},
                 ],
             }
         )
 
         funnel_trends = ClickhouseFunnelTrendsNew(filter, self.team)
         results = funnel_trends.run()
+
+        self.assertListEqual(
+            cast(list, results),
+            [
+                (datetime(2021, 5, 31, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 1, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 2, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 3, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 4, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 5, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 6, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 7, 0, 0), 0, 0, 0.0),
+            ],
+        )
+
+    def test_only_one_user_reached_one_step(self):
+        _create_person(distinct_ids=[USER_A_DISTINCT_ID], team=self.team)
+
+        _create_event(event=STEP_1_EVENT, distinct_id=USER_A_DISTINCT_ID, team=self.team, timestamp=TIME_3)
+
+        filter = Filter(
+            data={
+                "insight": INSIGHT_FUNNELS,
+                "display": TRENDS_LINEAR,
+                "interval": "day",
+                "date_from": TIME_1,
+                "date_to": TIME_99,
+                "funnel_window_days": 7,
+                "events": [
+                    {"id": STEP_1_EVENT, "order": 0},
+                    {"id": STEP_2_EVENT, "order": 1},
+                    {"id": STEP_3_EVENT, "order": 2},
+                ],
+            }
+        )
+
+        funnel_trends = ClickhouseFunnelTrendsNew(filter, self.team)
+        results = funnel_trends.run()
+
+        self.assertListEqual(
+            cast(list, results),
+            [
+                (datetime(2021, 5, 31, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 1, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 2, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 3, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 4, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 5, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 6, 0, 0), 0, 0, 0.0),
+                (datetime(2021, 6, 7, 0, 0), 1, 0, 0.0),
+            ],
+        )
